api/requirements.txt
fastapi[standard]
uvicorn[standard]
PyJWT
immudb-py
python-dotenv
passlib[bcrypt]
pydantic-settings